Change PlatformEffect back, default container to null
