+++ Xamarin.Forms.Platform.Android/FastRenderers/EffectControlProvider.cs
@@ -1,17 +1,26 @@
+using Android.Views;
 using AView = Android.Views.View;
 
 namespace Xamarin.Forms.Platform.Android.FastRenderers
 {
 	internal class EffectControlProvider : IEffectControlProvider
 	{
-		readonly AView _control; 
+		readonly AView _control;
+	    readonly ViewGroup _container;
 
 		public EffectControlProvider(AView control)
 		{
 			_control = control;
+		    _container = null;
 		}
 
-		public void RegisterEffect(Effect effect)
+        public EffectControlProvider(AView control, ViewGroup container)
+        {
+            _control = control;
+            _container = container;
+        }
+
+        public void RegisterEffect(Effect effect)
 		{
 			var platformEffect = effect as PlatformEffect;
 			if (platformEffect == null)
@@ -20,7 +29,7 @@ namespace Xamarin.Forms.Platform.Android.FastRenderers
 			}
 
 			platformEffect.SetControl(_control);
-			platformEffect.SetContainer(_control);
+			platformEffect.SetContainer(_container);
 		}
 	}
 }
