+++ Xamarin.Forms.Platform.Android/PlatformEffect.cs
@@ -4,7 +4,7 @@ using AView = Android.Views.View;
 
 namespace Xamarin.Forms.Platform.Android
 {
-	public abstract class PlatformEffect : PlatformEffect<AView, AView>
+	public abstract class PlatformEffect : PlatformEffect<ViewGroup, AView>
 	{
 	}
 }
